Refactor ui elements into separate package and modules
